+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - added: Add Monad (MON) support
+- added: Nym (NYM) Cosmos-based chain with NYX builtin token
+- added: (Ethereum) NYM ERC-20 builtin token
+- added: (BSC) NYM BEP-20 builtin token
 
 ## 4.71.2 (2026-02-02)